animation + dead zone + attempt on item pick up
implemented in the dead zone for the player and platform added textures to the animator but they are happening randomly (needs player control script).
also made I start to the item pick up with adding in the pickup script to the coconut object but I will still need help on that.